+++ packages/render/index.js
@@ -4,4 +4,5 @@ const engine = require('./engine');
 const site = require('./site');
 
 const documents = require('./documents');
-module.exports = { ...engine, ...site, documents };
+const dates = require('./dates');
+module.exports = { ...engine, ...site, ...dates, documents };